(TextEditor) Rename text editor next to text editor tiptap
